requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
black==19.10b0
cfgv==3.1.0
click==7.1.2
distlib==0.3.1
Django==3.1b1
django-cors-headers==3.4.0
djangorestframework==3.11.0
filelock==3.0.12
identify==1.4.21
importlib-metadata==1.7.0
importlib-resources==3.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nodeenv==1.4.0
pathspec==0.8.0
pre-commit==2.6.0
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
pytube3==9.6.4
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
virtualenv==20.0.25
wrapt==1.12.1
zipp==3.1.0
